requirements.txt
# Core dependencies
torch>=2.0.0
pytorch-lightning>=2.0.0
transformers>=4.30.0
tokenizers>=0.22.2  # Prevent downgrade that conflicts with antiberty

# Data processing
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Visualization and logging
wandb>=0.15.0
matplotlib>=3.5.0
seaborn>=0.13.0

# ML models
xgboost>=2.0.0

# Utilities
scipy>=1.10.0
pyyaml>=6.0
tqdm>=4.65.0

# Protein models
antiberty>=0.1.0   # Official AntiBERTy package
abnumber>=0.3.2    # Antibody numbering (IMGT/Chothia/Kabat)
biopython>=1.80    # Sequence I/O (Bio.SeqIO) and isoelectric point calculation
# biophi>=1.0.0    # Humanness scoring (OASis) - optional, install separately if needed
# ESM-2 is accessed via transformers
# ESM-C requires: pip install git+https://github.com/evolutionaryscale/esm.git
# ProtT5 is accessed via transformers
